feat(rpc): remove v0.2
