+++ crates/rpc/src/v02/method.rs
@@ -4,15 +4,12 @@ mod add_invoke_transaction;
 mod block_hash_and_number;
 pub(crate) mod call;
 mod chain_id;
-pub(crate) mod estimate_fee;
 mod get_block;
 mod get_block_transaction_count;
 mod get_class;
 mod get_class_at;
 mod get_class_hash_at;
-mod get_events;
 mod get_nonce;
-pub(crate) mod get_state_update;
 mod get_storage_at;
 mod get_transaction_by_block_id_and_index;
 mod get_transaction_by_hash;
@@ -26,15 +23,12 @@ pub(crate) use add_invoke_transaction::add_invoke_transaction;
 pub(crate) use block_hash_and_number::{block_hash_and_number, block_number};
 pub(crate) use call::call;
 pub(crate) use chain_id::chain_id;
-pub(super) use estimate_fee::estimate_fee;
 pub(crate) use get_block::{get_block_with_tx_hashes, get_block_with_txs};
 pub(crate) use get_block_transaction_count::get_block_transaction_count;
 pub(crate) use get_class::get_class;
 pub(crate) use get_class_at::get_class_at;
 pub(crate) use get_class_hash_at::get_class_hash_at;
-pub(super) use get_events::get_events;
 pub(crate) use get_nonce::get_nonce;
-pub(super) use get_state_update::get_state_update;
 pub(crate) use get_storage_at::get_storage_at;
 pub(crate) use get_transaction_by_block_id_and_index::get_transaction_by_block_id_and_index;
 pub(crate) use get_transaction_by_hash::get_transaction_by_hash;
