--- crates/rpc/src/v02/method/get_events.rs
@@ -1,883 +1,0 @@
-use crate::context::RpcContext;
-use anyhow::Context;
-use pathfinder_common::{BlockId, BlockNumber, ContractAddress, EventKey};
-use pathfinder_storage::{EventFilter, EventFilterError, V02KeyFilter};
-use serde::Deserialize;
-use starknet_gateway_types::reply::PendingBlock;
-use tokio::task::JoinHandle;
-
-crate::error::generate_rpc_error_subset!(
-    GetEventsError: BlockNotFound,
-    PageSizeTooBig,
-    InvalidContinuationToken
-);
-
-#[derive(serde::Deserialize, Debug, PartialEq, Eq)]
-#[cfg_attr(test, derive(Clone))]
-pub struct GetEventsInput {
-    filter: InputFilter,
-}
-
-/// Contains event filter parameters passed to `starknet_getEvents`.
-#[serde_with::skip_serializing_none]
-#[derive(Clone, Debug, Deserialize, PartialEq, Eq)]
-#[serde(deny_unknown_fields)]
-pub struct InputFilter {
-    #[serde(default)]
-    pub from_block: Option<BlockId>,
-    #[serde(default)]
-    pub to_block: Option<BlockId>,
-    #[serde(default)]
-    pub address: Option<ContractAddress>,
-    #[serde(default)]
-    pub keys: Vec<EventKey>,
-
-    // These are inlined here because serde flatten and deny_unknown_fields
-    // don't work together.
-    pub chunk_size: usize,
-    /// Offset, measured in events, which points to the requested chunk
-    #[serde(default)]
-    pub continuation_token: Option<String>,
-}
-
-/// Returns events matching the specified filter
-pub async fn get_events(
-    context: RpcContext,
-    input: GetEventsInput,
-) -> Result<types::GetEventsResult, GetEventsError> {
-    // The [Block::Pending] in ranges makes things quite complicated. This implementation splits
-    // the ranges into the following buckets:
-    //
-    // 1. pending     :     pending -> query pending only
-    // 2. pending     : non-pending -> return empty result
-    // 3. non-pending : non-pending -> query db only
-    // 4. non-pending :     pending -> query db and potentially append pending events
-    //
-    // The database query for 3 and 4 is combined into one step.
-    //
-    // 4 requires some additional logic to handle some edge cases:
-    //  a) Query database
-    //  b) if full page           -> return page
-    //  c) else if partially full -> append events from start of pending
-    //  d) else (page is empty):
-    //      i) query database for event count
-    //     ii) query pending data using count for paging into pending events
-
-    use BlockId::*;
-
-    let request = input.filter;
-    let requested_offset = match request.continuation_token {
-        Some(s) => Some(
-            s.parse::<usize>()
-                .map_err(|_| GetEventsError::InvalidContinuationToken)?,
-        ),
-        None => None,
-    };
-
-    // Grab the pending block so that we can check its validity.
-    //
-    // This is an async operation, so separating it from the sync database portion is sensible.
-    let pending_block = match context.pending_data.as_ref() {
-        Some(data) => data.block().await,
-        None => None,
-    };
-
-    // Handle the trivial (1) and (2) cases.
-    match (request.from_block, request.to_block) {
-        (Some(Pending), non_pending) if non_pending != Some(Pending) => {
-            return Ok(types::GetEventsResult {
-                events: Vec::new(),
-                continuation_token: None,
-            });
-        }
-        (Some(Pending), Some(Pending)) => {
-            let skip = requested_offset.unwrap_or_default();
-
-            let mut events = Vec::new();
-            let is_last_page = match pending_block {
-                None => true,
-                Some(pending_block) => {
-                    append_pending_events(
-                        &pending_block,
-                        &mut events,
-                        skip,
-                        request.chunk_size,
-                        request.address,
-                        request.keys.into_iter().collect(),
-                    )
-                    .await
-                }
-            };
-
-            check_continuation_token_validity(requested_offset, &events)?;
-
-            let continuation_token =
-                next_continuation_token(skip, request.chunk_size, is_last_page);
-
-            return Ok(types::GetEventsResult {
-                events,
-                continuation_token,
-            });
-        }
-        _ => {}
-    }
-
-    let storage = context.storage.clone();
-    let keys = V02KeyFilter::new(request.keys.clone());
-
-    // blocking task to perform database event query and optionally, the event count
-    // required for (4d).
-
-    let span = tracing::Span::current();
-    let db_events: JoinHandle<Result<_, GetEventsError>> = tokio::task::spawn_blocking(move || {
-        let _g = span.enter();
-        let mut connection = storage
-            .connection()
-            .context("Opening database connection")?;
-
-        let transaction = connection
-            .transaction()
-            .context("Creating database transaction")?;
-
-        let from_block = map_from_block_to_number(&transaction, request.from_block)?;
-        let to_block = map_to_block_to_number(&transaction, request.to_block)?;
-
-        let filter = EventFilter {
-            from_block,
-            to_block,
-            contract_address: request.address,
-            keys: keys.clone(),
-            page_size: request.chunk_size,
-            offset: requested_offset.unwrap_or_default(),
-        };
-        // We don't add context here, because [StarknetEventsTable::get_events] adds its
-        // own context to the errors. This way we get meaningful error information
-        // for errors related to query parameters.
-        let page = transaction.events(&filter).map_err(|e| -> GetEventsError {
-            if let Some(event_filter_error) = e.downcast_ref::<EventFilterError>() {
-                match event_filter_error {
-                    EventFilterError::PageSizeTooBig(_) => GetEventsError::PageSizeTooBig,
-                    EventFilterError::TooManyMatches => GetEventsError::from(e),
-                }
-            } else {
-                GetEventsError::from(e)
-            }
-        })?;
-
-        // Additional information is required if we need to append pending events.
-        // More specifically, we need some database event count in order to page through
-        // the pending events properly.
-        let event_count = if request.to_block == Some(Pending) && page.events.is_empty() {
-            let count = transaction.event_count(from_block, to_block, request.address, &keys)?;
-
-            Some(count)
-        } else {
-            None
-        };
-
-        let continuation_token =
-            next_continuation_token(filter.offset, filter.page_size, page.is_last_page);
-
-        // Check pending block validity if it is required -- this means checking it's parent is
-        // indeed the latest block in storage. Replace pending data if it is invalid, or not required.
-        let pending_block = match (request.to_block, pending_block) {
-            (Some(Pending), Some(pending_block)) => {
-                let latest = transaction
-                    .block_id(pathfinder_storage::BlockId::Latest)
-                    .context("Querying latest block hash")?
-                    .context("Latest block hash is missing")?
-                    .1;
-
-                (pending_block.parent_hash == latest).then_some(pending_block)
-            }
-            (_, original) => original,
-        };
-
-        Ok((
-            types::GetEventsResult {
-                events: page.events.into_iter().map(|e| e.into()).collect(),
-                continuation_token,
-            },
-            event_count,
-            pending_block,
-        ))
-    });
-
-    let (mut events, count, pending_block) = db_events
-        .await
-        .context("Database read panic or shutting down")??;
-
-    // Append pending data if required.
-    if matches!(request.to_block, Some(Pending))
-        && events.events.len() < request.chunk_size
-        && pending_block.is_some()
-    {
-        // This is safe as we just check pending_block.is_some() above.
-        let pending_block = pending_block.unwrap();
-
-        let keys = request
-            .keys
-            .into_iter()
-            .collect::<std::collections::HashSet<_>>();
-
-        let amount = request.chunk_size - events.events.len();
-
-        let skip = match count {
-            Some(count) => {
-                // This will not yield an underflow error, as when continuation_token is None,
-                // then the count is also always 0, since the last page can only be empty if there's
-                // only a single empty page
-                requested_offset.unwrap_or_default() - count
-            }
-            None => 0,
-        };
-
-        let is_last_page = append_pending_events(
-            &pending_block,
-            &mut events.events,
-            skip,
-            amount,
-            request.address,
-            keys,
-        )
-        .await;
-
-        events.continuation_token = next_continuation_token(
-            requested_offset.unwrap_or_default(),
-            request.chunk_size,
-            is_last_page,
-        );
-    }
-
-    check_continuation_token_validity(requested_offset, &events.events)?;
-
-    Ok(events)
-}
-
-// Maps `to_block` BlockId to a block number which can be used by the events query.
-//
-// This block id specifies the upper end of the range, so pending/latest/None means
-// there's no upper limit.
-fn map_to_block_to_number(
-    tx: &pathfinder_storage::Transaction<'_>,
-    block: Option<BlockId>,
-) -> Result<Option<BlockNumber>, GetEventsError> {
-    use BlockId::*;
-
-    match block {
-        Some(Hash(hash)) => {
-            let number = tx
-                .block_id(pathfinder_storage::BlockId::Hash(hash))?
-                .ok_or(GetEventsError::BlockNotFound)?
-                .0;
-
-            Ok(Some(number))
-        }
-        Some(Number(number)) => Ok(Some(number)),
-        Some(Pending) | Some(Latest) | None => Ok(None),
-    }
-}
-
-// Maps `from_block` BlockId to a block number which can be used by the events query.
-//
-// This block id specifies the lower end of the range, so pending/latest means
-// a lower limit here.
-fn map_from_block_to_number(
-    tx: &pathfinder_storage::Transaction<'_>,
-    block: Option<BlockId>,
-) -> Result<Option<BlockNumber>, GetEventsError> {
-    use BlockId::*;
-
-    match block {
-        Some(Hash(hash)) => {
-            let number = tx
-                .block_id(pathfinder_storage::BlockId::Hash(hash))?
-                .ok_or(GetEventsError::BlockNotFound)?
-                .0;
-
-            Ok(Some(number))
-        }
-        Some(Number(number)) => Ok(Some(number)),
-        Some(Pending) | Some(Latest) => {
-            let number = tx
-                .block_id(pathfinder_storage::BlockId::Latest)?
-                .ok_or(GetEventsError::BlockNotFound)?
-                .0;
-            Ok(Some(number))
-        }
-        None => Ok(None),
-    }
-}
-
-/// Append's pending events to `dst` based on the filter requirements and returns
-/// true if this was the last pending data i.e. `is_last_page`.
-async fn append_pending_events(
-    pending_block: &PendingBlock,
-    dst: &mut Vec<types::EmittedEvent>,
-    skip: usize,
-    amount: usize,
-    address: Option<ContractAddress>,
-    keys: std::collections::HashSet<EventKey>,
-) -> bool {
-    let original_len = dst.len();
-
-    let pending_events = pending_block
-        .transaction_receipts
-        .iter()
-        .flat_map(|receipt| {
-            receipt
-                .events
-                .iter()
-                .zip(std::iter::repeat(receipt.transaction_hash))
-        })
-        .filter(|(event, _)| match address {
-            Some(address) => event.from_address == address,
-            None => true,
-        })
-        .filter(|(event, _)| {
-            if keys.is_empty() {
-                return true;
-            }
-
-            for ek in &event.keys {
-                if keys.contains(ek) {
-                    return true;
-                }
-            }
-            false
-        })
-        .skip(skip)
-        // We need to take an extra event to determine is_last_page.
-        .take(amount + 1)
-        .map(|(event, tx_hash)| types::EmittedEvent {
-            data: event.data.clone(),
-            keys: event.keys.clone(),
-            from_address: event.from_address,
-            block_hash: None,
-            block_number: None,
-            transaction_hash: tx_hash,
-        });
-
-    dst.extend(pending_events);
-    let is_last_page = dst.len() <= (original_len + amount);
-    if !is_last_page {
-        dst.pop();
-    }
-
-    is_last_page
-}
-
-/// Continuation token is invalid if:
-/// 1. its value is nonzero (which means that it _actually_ points to some _next page_)
-/// 2. it yields an empty page
-///
-/// Unfortunately page retrieval has to be completed before the actual check can be done.
-fn check_continuation_token_validity(
-    continuation_token: Option<usize>,
-    events: &[types::EmittedEvent],
-) -> Result<(), GetEventsError> {
-    match continuation_token {
-        Some(token) if token > 0 && events.is_empty() => {
-            Err(GetEventsError::InvalidContinuationToken)
-        }
-        Some(_) | None => Ok(()),
-    }
-}
-
-fn next_continuation_token(
-    current_offset: usize,
-    chunk_size: usize,
-    is_last_page: bool,
-) -> Option<String> {
-    if is_last_page {
-        None
-    } else {
-        // Point to the next page
-        Some((current_offset + chunk_size).to_string())
-    }
-}
-
-mod types {
-    use crate::felt::{RpcFelt, RpcFelt251};
-    use pathfinder_common::{
-        BlockHash, BlockNumber, ContractAddress, EventData, EventKey, TransactionHash,
-    };
-    use serde::Serialize;
-
-    /// Describes an emitted event returned by starknet_getEvents
-    #[serde_with::serde_as]
-    #[derive(Clone, Debug, Serialize, PartialEq, Eq)]
-    #[serde(deny_unknown_fields)]
-    pub struct EmittedEvent {
-        #[serde_as(as = "Vec<RpcFelt>")]
-        pub data: Vec<EventData>,
-        #[serde_as(as = "Vec<RpcFelt>")]
-        pub keys: Vec<EventKey>,
-        #[serde_as(as = "RpcFelt251")]
-        pub from_address: ContractAddress,
-        /// [None] for pending events.
-        #[serde_as(as = "Option<RpcFelt>")]
-        pub block_hash: Option<BlockHash>,
-        /// [None] for pending events.
-        pub block_number: Option<BlockNumber>,
-        #[serde_as(as = "RpcFelt")]
-        pub transaction_hash: TransactionHash,
-    }
-
-    impl From<pathfinder_storage::EmittedEvent> for EmittedEvent {
-        fn from(event: pathfinder_storage::EmittedEvent) -> Self {
-            Self {
-                data: event.data,
-                keys: event.keys,
-                from_address: event.from_address,
-                block_hash: Some(event.block_hash),
-                block_number: Some(event.block_number),
-                transaction_hash: event.transaction_hash,
-            }
-        }
-    }
-
-    // Result type for starknet_getEvents
-    #[serde_with::skip_serializing_none]
-    #[derive(Clone, Debug, Serialize, PartialEq, Eq)]
-    #[serde(deny_unknown_fields)]
-    pub struct GetEventsResult {
-        pub events: Vec<EmittedEvent>,
-        /// Offset, measured in events, which points to the chunk that follows currenty requested chunk (`events`)
-        pub continuation_token: Option<String>,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{
-        types::{EmittedEvent, GetEventsResult},
-        *,
-    };
-    use jsonrpsee::types::Params;
-    use pathfinder_common::felt;
-    use pathfinder_common::macro_prelude::*;
-    use pathfinder_storage::test_utils;
-    use pretty_assertions::assert_eq;
-
-    #[test]
-    fn parsing() {
-        let optional_present = InputFilter {
-            from_block: Some(BlockId::Number(BlockNumber::new_or_panic(0))),
-            to_block: Some(BlockId::Latest),
-            address: Some(contract_address!("0x1")),
-            keys: vec![event_key!("0x2")],
-            chunk_size: 3,
-            continuation_token: Some("4".to_string()),
-        };
-        let optional_absent = InputFilter {
-            from_block: None,
-            to_block: None,
-            address: None,
-            keys: vec![],
-            chunk_size: 5,
-            continuation_token: None,
-        };
-
-        [
-            (
-                r#"[{"from_block":{"block_number":0},"to_block":"latest","address":"0x1","keys":["0x2"],"chunk_size":3,"continuation_token":"4"}]"#,
-                optional_present.clone(),
-            ),
-            (
-                r#"{"filter":{"from_block":{"block_number":0},"to_block":"latest","address":"0x1","keys":["0x2"],"chunk_size":3,"continuation_token":"4"}}"#,
-                optional_present
-            ),
-            (r#"[{"chunk_size":5}]"#, optional_absent.clone()),
-            (r#"{"filter":{"chunk_size":5}}"#, optional_absent),
-        ]
-        .into_iter()
-        .enumerate()
-        .for_each(|(i, (input, expected))| {
-            let actual = Params::new(Some(input))
-                .parse::<GetEventsInput>()
-                .unwrap_or_else(|error| panic!("test case {i}: {input}, {error}"));
-            assert_eq!(
-                actual,
-                GetEventsInput { filter: expected },
-                "test case {i}: {input}"
-            );
-        });
-    }
-
-    fn setup() -> (RpcContext, Vec<EmittedEvent>) {
-        let (storage, test_data) = test_utils::setup_test_storage();
-        let events = test_data
-            .events
-            .into_iter()
-            .map(EmittedEvent::from)
-            .collect();
-        let context = RpcContext::for_tests().with_storage(storage);
-
-        (context, events)
-    }
-
-    impl PartialEq for GetEventsError {
-        fn eq(&self, other: &Self) -> bool {
-            match (self, other) {
-                (Self::Internal(l), Self::Internal(r)) => l.to_string() == r.to_string(),
-                _ => core::mem::discriminant(self) == core::mem::discriminant(other),
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn get_events_with_empty_filter() {
-        let (context, events) = setup();
-
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: None,
-                to_block: None,
-                address: None,
-                keys: vec![],
-                chunk_size: test_utils::NUM_EVENTS,
-                continuation_token: None,
-            },
-        };
-        let result = get_events(context, input).await.unwrap();
-
-        assert_eq!(
-            result,
-            GetEventsResult {
-                events,
-                continuation_token: None,
-            }
-        );
-    }
-
-    #[tokio::test]
-    async fn get_events_with_fully_specified_filter() {
-        let (context, events) = setup();
-
-        let expected_event = &events[1];
-        let expected_result = GetEventsResult {
-            events: vec![expected_event.clone()],
-            continuation_token: None,
-        };
-        let mut input = GetEventsInput {
-            filter: InputFilter {
-                from_block: Some(expected_event.block_number.unwrap().into()),
-                to_block: Some(expected_event.block_number.unwrap().into()),
-                address: Some(expected_event.from_address),
-                // we're using a key which is present in _all_ events
-                keys: vec![event_key!("0xdeadbeef")],
-                chunk_size: test_utils::NUM_EVENTS,
-                continuation_token: None,
-            },
-        };
-        let result = get_events(context.clone(), input.clone()).await.unwrap();
-        assert_eq!(result, expected_result);
-
-        // 0 continuation token should yield the same result as no token
-        input.filter.continuation_token = Some(0.to_string());
-        let result = get_events(context, input).await.unwrap();
-        assert_eq!(result, expected_result);
-    }
-
-    #[tokio::test]
-    async fn get_events_by_block() {
-        let (context, events) = setup();
-
-        const BLOCK_NUMBER: usize = 2;
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: Some(BlockNumber::new_or_panic(BLOCK_NUMBER as u64).into()),
-                to_block: Some(BlockNumber::new_or_panic(BLOCK_NUMBER as u64).into()),
-                address: None,
-                keys: vec![],
-                chunk_size: test_utils::NUM_EVENTS,
-                continuation_token: None,
-            },
-        };
-
-        let result = get_events(context, input).await.unwrap();
-
-        let expected_events = &events[test_utils::EVENTS_PER_BLOCK * BLOCK_NUMBER
-            ..test_utils::EVENTS_PER_BLOCK * (BLOCK_NUMBER + 1)];
-        assert_eq!(
-            result,
-            GetEventsResult {
-                events: expected_events.to_vec(),
-                continuation_token: None,
-            }
-        );
-    }
-
-    #[tokio::test]
-    async fn get_events_from_latest_block() {
-        let (context, events) = setup();
-
-        const LATEST_BLOCK_NUMBER: usize = 3;
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: Some(BlockId::Latest),
-                to_block: Some(BlockId::Latest),
-                address: None,
-                keys: vec![],
-                chunk_size: test_utils::NUM_EVENTS,
-                continuation_token: None,
-            },
-        };
-
-        let result = get_events(context, input).await.unwrap();
-
-        let expected_events = &events[test_utils::EVENTS_PER_BLOCK * LATEST_BLOCK_NUMBER
-            ..test_utils::EVENTS_PER_BLOCK * (LATEST_BLOCK_NUMBER + 1)];
-        assert_eq!(
-            result,
-            GetEventsResult {
-                events: expected_events.to_vec(),
-                continuation_token: None,
-            }
-        );
-    }
-
-    #[tokio::test]
-    async fn get_events_with_invalid_page_size() {
-        let (context, _) = setup();
-
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: None,
-                to_block: None,
-                address: None,
-                keys: vec![],
-                chunk_size: pathfinder_storage::PAGE_SIZE_LIMIT + 1,
-                continuation_token: None,
-            },
-        };
-        let error = get_events(context, input).await.unwrap_err();
-
-        assert_eq!(GetEventsError::PageSizeTooBig, error);
-    }
-
-    #[tokio::test]
-    async fn get_events_by_key_with_paging() {
-        let (context, events) = setup();
-
-        let expected_events = &events[27..33];
-        let keys_for_expected_events: Vec<_> = expected_events.iter().map(|e| e.keys[0]).collect();
-
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: None,
-                to_block: None,
-                address: None,
-                keys: keys_for_expected_events.clone(),
-                chunk_size: 1,
-                continuation_token: None,
-            },
-        };
-        let result = get_events(context.clone(), input).await.unwrap();
-        assert_eq!(
-            result,
-            GetEventsResult {
-                events: expected_events[..1].to_vec(),
-                continuation_token: Some(1.to_string()),
-            }
-        );
-
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: None,
-                to_block: None,
-                address: None,
-                keys: keys_for_expected_events.clone(),
-                chunk_size: 2,
-                continuation_token: Some(1.to_string()),
-            },
-        };
-        let result = get_events(context.clone(), input).await.unwrap();
-        assert_eq!(
-            result,
-            GetEventsResult {
-                events: expected_events[1..3].to_vec(),
-                continuation_token: Some(3.to_string()),
-            }
-        );
-
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: None,
-                to_block: None,
-                address: None,
-                keys: keys_for_expected_events.clone(),
-                chunk_size: 3,
-                continuation_token: Some(3.to_string()),
-            },
-        };
-        let result = get_events(context.clone(), input).await.unwrap();
-        assert_eq!(
-            result,
-            GetEventsResult {
-                events: expected_events[3..].to_vec(),
-                continuation_token: None,
-            }
-        );
-
-        // nonexistent page
-        let input = GetEventsInput {
-            filter: InputFilter {
-                from_block: None,
-                to_block: None,
-                address: None,
-                keys: keys_for_expected_events.clone(),
-                chunk_size: 1,
-                // Offset pointing to after the last event
-                continuation_token: Some(6.to_string()),
-            },
-        };
-        let error = get_events(context, input).await.unwrap_err();
-        assert_eq!(error, GetEventsError::InvalidContinuationToken);
-    }
-
-    mod pending {
-        use super::*;
-        use pretty_assertions::assert_eq;
-        use starknet_gateway_types::pending::PendingData;
-
-        #[tokio::test]
-        async fn backward_range() {
-            let context = RpcContext::for_tests_with_pending().await;
-
-            let input = GetEventsInput {
-                filter: InputFilter {
-                    from_block: Some(BlockId::Pending),
-                    to_block: Some(BlockId::Latest),
-                    address: None,
-                    keys: vec![],
-                    chunk_size: 100,
-                    continuation_token: None,
-                },
-            };
-            let result = get_events(context, input).await.unwrap();
-            assert!(result.events.is_empty());
-        }
-
-        #[tokio::test]
-        async fn all_events() {
-            let context = RpcContext::for_tests_with_pending().await;
-
-            let input0 = GetEventsInput {
-                filter: InputFilter {
-                    from_block: None,
-                    to_block: Some(BlockId::Latest),
-                    address: None,
-                    keys: vec![],
-                    chunk_size: 1024,
-                    continuation_token: None,
-                },
-            };
-            let mut input1 = input0.clone();
-            input1.filter.continuation_token = Some(0.to_string());
-
-            for mut input in [input0, input1] {
-                let events = get_events(context.clone(), input.clone()).await.unwrap();
-
-                input.filter.from_block = Some(BlockId::Pending);
-                input.filter.to_block = Some(BlockId::Pending);
-                let pending_events = get_events(context.clone(), input.clone()).await.unwrap();
-
-                input.filter.from_block = None;
-                let all_events = get_events(context.clone(), input.clone()).await.unwrap();
-
-                let expected = events
-                    .events
-                    .into_iter()
-                    .chain(pending_events.events.into_iter())
-                    .collect::<Vec<_>>();
-
-                assert_eq!(all_events.events, expected);
-                assert!(all_events.continuation_token.is_none());
-            }
-        }
-
-        #[tokio::test]
-        async fn paging() {
-            let context = RpcContext::for_tests_with_pending().await;
-
-            let mut input = GetEventsInput {
-                filter: InputFilter {
-                    from_block: None,
-                    to_block: Some(BlockId::Pending),
-                    address: None,
-                    keys: vec![],
-                    chunk_size: 1024,
-                    continuation_token: None,
-                },
-            };
-
-            let all = get_events(context.clone(), input.clone())
-                .await
-                .unwrap()
-                .events;
-
-            input.filter.chunk_size = 2;
-            input.filter.continuation_token = Some(0.to_string()); // Should yield the same result as None above
-            let result = get_events(context.clone(), input.clone()).await.unwrap();
-            assert_eq!(result.events, &all[0..2]);
-            assert_eq!(result.continuation_token, Some(2.to_string()));
-
-            input.filter.chunk_size = 1;
-            input.filter.continuation_token = result.continuation_token;
-            let result = get_events(context.clone(), input.clone()).await.unwrap();
-            assert_eq!(result.events, &all[2..3]);
-            assert_eq!(result.continuation_token, Some(3.to_string()));
-
-            input.filter.chunk_size = 100; // Only a single event remains though
-            input.filter.continuation_token = result.continuation_token;
-            let result = get_events(context.clone(), input.clone()).await.unwrap();
-            assert_eq!(result.events, &all[3..4]);
-            assert_eq!(result.continuation_token, None);
-
-            // nonexistent page
-            input.filter.chunk_size = 123; // Does not matter
-            input.filter.continuation_token = Some(4.to_string()); // Points to after the last event
-            let error = get_events(context.clone(), input).await.unwrap_err();
-            assert_eq!(error, GetEventsError::InvalidContinuationToken);
-        }
-
-        #[tokio::test]
-        async fn invalid_pending() {
-            use std::sync::Arc;
-            // Invalidate the pending block data by changing its parent hash.
-            let mut context = RpcContext::for_tests_with_pending().await;
-            let pending = context.pending_data.as_ref().unwrap();
-            let mut block = pending.block().await.unwrap().as_ref().clone();
-            let state_update = pending.state_update().await.unwrap();
-            block.parent_hash.0 = block.parent_hash.0 + felt!("0x1");
-
-            let pending = PendingData::default();
-            pending.set(Arc::new(block), state_update).await;
-            context.pending_data = Some(pending);
-
-            let all_non_pending_filter = GetEventsInput {
-                filter: InputFilter {
-                    from_block: None,
-                    to_block: Some(BlockId::Latest),
-                    address: None,
-                    keys: vec![],
-                    chunk_size: 1024,
-                    continuation_token: None,
-                },
-            };
-
-            let mut all_filter = all_non_pending_filter.clone();
-            all_filter.filter.to_block = Some(BlockId::Pending);
-
-            let expected = get_events(context.clone(), all_non_pending_filter)
-                .await
-                .unwrap();
-
-            let all_events = get_events(context.clone(), all_filter).await.unwrap();
-
-            assert_eq!(all_events, expected);
-        }
-    }
-}
